pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "income-desk"
version = "0.3.1"
description = "Systematic options trading intelligence for small accounts. Regime detection, profitability gates, Kelly sizing, crash sentinel, decision audit. Your personal trading desk brain."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = [
    "options-trading", "iron-condor", "credit-spread", "theta",
    "regime-detection", "hmm", "kelly-criterion", "systematic-trading",
    "income-trading", "small-account", "risk-management",
    "tastytrade", "alpaca", "interactive-brokers", "schwab", "zerodha", "dhan",
    "options", "trading", "quant", "finance",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Typing :: Typed",
]
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "pydantic>=2.0",
    "hmmlearn>=0.3",
    "scikit-learn>=1.3",
    "joblib>=1.3",
    "yfinance>=0.2",
    "pyarrow>=14.0",
    "requests>=2.31",
    "tabulate>=0.9",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    # Broker SDKs — all included, any that fail to install are handled at runtime
    "tastytrade>=12.0",
    "alpaca-py>=0.30",
    "fredapi>=0.5",
    "dhanhq>=2.0",
]

[project.optional-dependencies]
# Brokers that have complex/platform-specific deps
ibkr = ["ib_insync>=0.9"]
schwab = ["schwab-py>=1.0"]
zerodha = ["kiteconnect>=5.0"]
dhan = ["dhanhq>=2.0"]
plot = ["matplotlib>=3.5"]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]
all = [
    "income-desk[ibkr,schwab,zerodha,dhan,plot]",
]

[project.scripts]
# Primary CLI
income-desk = "income_desk.cli.interactive:main"
# Legacy aliases (backward compat)
analyzer-cli = "income_desk.cli.interactive:main"
analyzer-explore = "income_desk.cli.explore:main"
analyzer-plot = "income_desk.cli.plot:main"
analyzer-rank = "income_desk.cli.rank:main"

[tool.hatch.build.targets.wheel]
packages = ["income_desk"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that hit real network APIs",
    "daily: marks tests that validate daily trading pipeline readiness",
]
